agentic_clone/requirements.txt
gradio
openai
python-dotenv
requests
pypdf
